Update to follow extended paper

+++ regular_pattern_test_classic.py
@@ -4,8 +4,8 @@ Created on Wed Nov 16 09:58:06 2022
 
 @author: Mieszko Ferens
 
-Script to run an experiment for modelling an XOR PUF that uses Binary-coded
-with Padding (BP) CRPs patterns as CRPs during authentication with the server
+Script to run an experiment for modelling an Arbiter PUF that uses regular
+patterns (BP CRPs) as CRPs during authentication with the server.
 """
 
 import argparse
@@ -23,7 +23,7 @@ class ChallengeResponseSet():
         self.responses = np.expand_dims(
             np.expand_dims(responses,axis=1),axis=1)
 
-def create_binary_code_challenges(n):
+def create_binary_code_challenges(n, N):
     
     n_bits = 16
     lsb = np.arange(2**n_bits, dtype=np.uint8).reshape(-1,1)
@@ -42,6 +42,13 @@ def create_binary_code_challenges(n):
     for i in range(1, int(n/n_bits)):
         challenges = np.append(challenges, np.roll(shift, i, axis=1), axis=0)
     
+    _ , idx = np.unique(challenges, return_index=True, axis=0)
+    challenges = challenges[np.sort(idx)]
+    
+    assert N <= len(challenges), (
+        "Not enough CRPs have been generated. The limit is (2^18 - 3) CRPs.")
+    challenges = challenges[:N]
+    
     return challenges
 
 def main():
@@ -56,31 +63,29 @@ def main():
     parser.add_argument("--seed", type=int, default=0, help="Random seed.")
     parser.add_argument("--n-bits", type=int, default=64,
                         help="Challenge length in bits.")
-    parser.add_argument("--train-data", type=int, default=160000,
+    parser.add_argument("--k", type=int, default=1,
+                        help="The number of parallel APUF in the XOR PUF.")
+    parser.add_argument("--train-data", type=int, default=80000,
                         help="Number of training data samples for the model.")
     parser.add_argument("--test-data", type=int, default=10000,
                         help="Number of testing data samples for the model.")
+    parser.add_argument("--ML-algorithm", type=str, default="LR",
+                        help="ML algorithm to model the PUF with (LR or MLP).")
     args = parser.parse_args()
     
     # Generate the PUF
-    k = 5
-    puf = XORArbiterPUF(args.n_bits, k, args.seed)
+    puf = XORArbiterPUF(args.n_bits, args.k, args.seed)
     
     # Generate the challenges
-    challenges = create_binary_code_challenges(n=args.n_bits)
-    
-    # Remove duplicate challenges
-    _ , idx = np.unique(challenges, return_index=True, axis=0)
-    challenges = challenges[np.sort(idx)]
-    
-    # Check if split the data into training and testing is possible
-    assert args.train_data + args.test_data <= len(challenges), (
-        "Not enough training data")
-    train = args.train_data
-    test = train + args.test_data
+    challenges = create_binary_code_challenges(
+        n=args.n_bits, N=(args.train_data + args.test_data))
     
     # Get responses
     responses = puf.eval(challenges)
+    
+    # Split the data into training and testing
+    train = args.train_data
+    test = train + args.test_data
     
     # Prepare the data for training and testing
     train_crps = ChallengeResponseSet(
@@ -89,22 +94,27 @@ def main():
     test_x = challenges[train:test]
     test_y = np.expand_dims(0.5 - 0.5*responses[train:test], axis=1)
 
-    # Use an MLP as a predictor
-    network = [16, 32, 16] # As defined in the literature: [2^(k-1), 2^k, 2^(k-1)]
-    model = pypuf.attack.MLPAttack2021(
-        train_crps, seed=args.seed, net=network, epochs=30, lr=.001, bs=1000,
-        early_stop=.08)
-    # model = pypuf.attack.LRAttack2021(
-    #     train_crps, seed=args.seed, k=k, epochs=100, lr=.001, bs=1000,
-    #     stop_validation_accuracy=.97)
+    if(args.ML_algorithm == "LR"): # Use LR as a predictor
+        model = pypuf.attack.LRAttack2021(
+            train_crps, seed=args.seed, k=args.k, epochs=100, lr=.001, bs=1000,
+            stop_validation_accuracy=.97)
+    elif(args.ML_algorithm == "MLP"): # Use MLP as a predictor
+        if(args.k <= 4): # If the XOR PUF is small don't reduce the NN too much
+            network = [8, 16, 8] 
+        else: # As defined in the literature: [2^(k-1), 2^k, 2^(k-1)]
+            network = [2**(args.k-1), 2**args.k, 2**(args.k-1)]
+        model = pypuf.attack.MLPAttack2021(
+            train_crps, seed=args.seed, net=network, epochs=30, lr=.001,
+            bs=1000, early_stop=.08)
+    else:
+        raise NotImplementedError("Only LR and MLP are supported.")
     
     # Train the model
     model.fit()
 
     # Test the model
     pred_y = model._model.eval(test_x)
-    
-    # pred_y = pred_y.reshape(len(pred_y), 1) # TODO (for LR output only)
+    pred_y = pred_y.reshape(len(pred_y), 1)
     
     # Calculate accuracy
     accuracy = np.count_nonzero(((pred_y<0.5) + test_y)-1)/len(test_y)
@@ -112,12 +122,20 @@ def main():
           "Accuracy in the testing data: " + str(accuracy*100) + "%")
     
     # Log data into csv format
-    filepath = Path(args.outdir + "out_regular.csv")
-    filepath.parent.mkdir(parents=True, exist_ok=True)
     data = pd.DataFrame({"seed": [args.seed],
+                         "n_bits": [args.n_bits],
+                         "k": [args.k],
                          "train_data": [args.train_data],
+                         "test_data": [args.test_data],
+                         "ML_algorithm": [args.ML_algorithm],
                          "accuracy": [accuracy]})
-    data.to_csv(filepath, header=False, index=False, mode='a')
+    filepath = Path(args.outdir + "out_regular_pattern_" + str(args.k) +
+                    "XOR.csv")
+    if(filepath.is_file()):
+        data.to_csv(filepath, header=False, index=False, mode='a')
+    else:
+        filepath.parent.mkdir(parents=True, exist_ok=True)
+        data.to_csv(filepath, header=True, index=False, mode='a')
     
 
 if(__name__ == "__main__"):
